+++ timeutil/timeutil_test.go
@@ -53,8 +53,8 @@ func (s *TimeUtilSuite) TestPretyDuration(c *C) {
 
 func (s *TimeUtilSuite) TestPretyDurationInDays(c *C) {
 	c.Assert(PrettyDurationInDays("ABC"), Equals, "")
-	c.Assert(PrettyDurationInDays(120), Equals, "just now")
-	c.Assert(PrettyDurationInDays(7200), Equals, "today")
+	c.Assert(PrettyDurationInDays(120), Equals, "1 day")
+	c.Assert(PrettyDurationInDays(7200), Equals, "1 day")
 	c.Assert(PrettyDurationInDays(90000), Equals, "1 day")
 	c.Assert(PrettyDurationInDays(1296000), Equals, "15 days")
 }
@@ -289,6 +289,14 @@ func (s *TimeUtilSuite) TestHelpers(c *C) {
 	c.Assert(PrevWeekend(d), DeepEquals, time.Date(2021, 7, 31, 12, 30, 15, 0, time.Local))
 	c.Assert(NextWorkday(d), DeepEquals, time.Date(2021, 8, 2, 12, 30, 15, 0, time.Local))
 	c.Assert(NextWeekend(d), DeepEquals, time.Date(2021, 8, 7, 12, 30, 15, 0, time.Local))
+
+	d = time.Date(2021, 8, 13, 12, 30, 15, 0, time.Local)
+	c.Assert(StartOfHour(d), DeepEquals, time.Date(2021, 8, 13, 12, 0, 0, 0, time.Local))
+	c.Assert(StartOfDay(d), DeepEquals, time.Date(2021, 8, 13, 0, 0, 0, 0, time.Local))
+	c.Assert(StartOfWeek(d, time.Monday), DeepEquals, time.Date(2021, 8, 9, 0, 0, 0, 0, time.Local))
+	c.Assert(StartOfWeek(time.Time{}, time.Monday), DeepEquals, time.Time{})
+	c.Assert(StartOfMonth(d), DeepEquals, time.Date(2021, 8, 1, 0, 0, 0, 0, time.Local))
+	c.Assert(StartOfYear(d), DeepEquals, time.Date(2021, 1, 1, 0, 0, 0, 0, time.Local))
 }
 
 // ////////////////////////////////////////////////////////////////////////////////// //