+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 12.54.0
+
+* `[timeutil]` Added more helpers for working with dates
+* `[options]` Fixed panic in `GetS` when mixed or string option contains non-string value
+
 ### 12.53.0
 
 * `[strutil]` Added method `B` for choosing value by condition